src/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.3.0
bs4==0.0.2
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==42.0.5
dataclasses-json==0.6.4
defusedxml==0.7.1
Django==4.2.10
django-allauth==0.61.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-environ==0.11.2
django-htmx==1.17.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-social-auth==0.7.28
django-timezone-field==6.1.0
django-widget-tweaks==1.5.0
dnspython==2.6.1
email_validator==2.1.1
frozenlist==1.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.5
langchain==0.1.15
langchain-community==0.0.32
langchain-core==0.1.41
langchain-google-genai==1.0.1
langchain-text-splitters==0.0.1
langsmith==0.1.42
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauth2==1.9.0.post1
oauthlib==3.2.2
orjson==3.10.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyparsing==3.1.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-openid==2.2.5
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
stripe==4.2.0
tenacity==8.2.3
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wheel==0.41.2
yarl==1.9.4
